+++ frontend/www/js/omegaup/grader/GoldenLayoutConfigs.ts
