+++ frontend/www/js/omegaup/grader/MonacoEditor.vue
