+++ frontend/www/js/omegaup/grader/TextEditor.vue
