+++ webpack.config-frontend.js
@@ -69,7 +69,7 @@ module.exports = {
     course_submissions_list:
       './frontend/www/js/omegaup/course/submissions_list.ts',
     course_tabs: './frontend/www/js/omegaup/course/tabs.ts',
-    grader_ide: './frontend/www/js/omegaup/graderv2/ide.ts',
+    grader_ide: './frontend/www/js/omegaup/grader/ide.ts',
     group_edit: './frontend/www/js/omegaup/group/edit.ts',
     group_list: './frontend/www/js/omegaup/group/list.ts',
     group_new: './frontend/www/js/omegaup/group/new.ts',
